Create FindAndEnforceSmartcardLogonRequired.ps1
Because manually checking the boxes for all accounts that are unenforced is not a good use of one's time when there is more than a single account. And because this finds all of them in one go, as well. And it can exclude accounts that are not able to have the SLR bit set.

Generally, just don't do anything manually that can be automated.