Make the DNS suffix obtaining code work as intended
Previously, the code obtaining the DNS suffix would only execute if the OU was not specified in the ``$ou`` variable at the top of the script. For testing, that works, but when actually pre-loading the OU via the variable, it does not work as intended. [Compare: "fooworkstation1." ( when ``$ou`` is set) versus "fooworkstation1.contoso.com" (when ``$ou`` is not set)]