Add API endpoints for getting/editing asset lists. (#11)
* Add API endpoints for getting/editing asset lists.

Also fix nesting bug in the POST `SC-Connect` path.

* Create Auto Update AD DNS Asset List.ps1

Add a new script to automatically update a specified SecurityCenter DNS asset list with system names pulled from ActiveDirectory

* Integrate 'Auto Update AD DNS Asset List'

Add the AD Asset list update script into the main script.

* Fix conditional statement

Only evaluate the block if needed (if => elseif)

* Make the DNS suffix obtaining code work as intended

Previously, the code obtaining the DNS suffix would only execute if the OU was not specified in the ``$ou`` variable at the top of the script. For testing, that works, but when actually pre-loading the OU via the variable, it does not work as intended. [Compare: "fooworkstation1." ( when ``$ou`` is set) versus "fooworkstation1.contoso.com" (when ``$ou`` is not set)]