Don't error when zero new IAVMs for selected period
Fix an issue discovered in production. Not a severe issue, but if zero new IAVMs were released in the selected span of time, an error would result when trying to generate the temporary file with the IAVM IDs. So shift around the code to only ensure we attempt doing that if there was at least one new IAVM released.