Use a stream reader instead of Get-Content
Get-Content works fine for small files, but it seriously chokes when you're trying to ingest anything of any relative size (unsure as to what the magic number is, but a respectably sized Nessus file for 500 IPs (900 MiB) will cause PoSH to consume an inordinate amount of RAM, and seemingly hang (or maybe it was paging to disk; no idea) to the point where I grew bored of waiting for it and just ended the process). This works without consuming crazy amounts of RAM, since we are essentially streaming the file from the disk, and only getting what we need from the file.