Fixes & Improvements (#15)
* Logic fixes

* Shunt logout information to /dev/null

We don't care about it. Hide it.

* Spelling fixes

I have no idea why L94-103 changed... it's a comment, anyway, so... ¯\_(ツ)_/¯

* Enhancements; dump more information.

* Permit selection from all available repositories

AKA, I am tired of needing to use the stand-alone script for repositories other than the main one.

The user is still ultimately responsible for selecting a repository the purge file can be imported to, of course.