Fix logging in with username/password
This is what happens when you never test a function. It just so happens that I've been intending to log in via PKI all this time, so this got missed.

As for the _reason_ this got missed, turns out something like ``[string]$pkiThumbprint = $null,`` will result in the variable always being a zero-value string. So it's never going to be ``$null``, no matter how you look at it.