No, really, ID always comes back.
https://docs.tenable.com/sccv/api/Scan-Zone.html

Not specifying anything in the field list results in everything being returned. So there's no need to manually specify the ID here (may, or may not decide to add this in as a separate switch, but for now, I've never _just_ needed the ID number of a zone and nothing else previously. So everything it is.