Create extract_plugins_from_sc_dvl.ps1 (#19)
Because while PoSH _can_ ingest CSVs, depending on the size Import-Csv might just cause the shell to fall over from the raw quantity of data being pulled in. So let's make something to just extract what we need.

* Create extract_plugins_from_sc_dvl.ps1

Because sometimes Excel falls over when trying to handle a 10GiB+ CSV, and you only need a few plugins from the mass CSV.

* Update extract_plugins_from_sc_dvl.ps1

Enhancements. Because why not.