Use [String]::Join instead of loop-join
Because why reinvent the wheel and use a foreach loop to manually join the string to a concatenated string? Let's use what the language gives us.

Execution time for this section reduced basically to O(1) instead of O(n), or whatever it was, now that we don't need to iterate over every IP in a CIDR/range ourselves.