Create Update DNS SCAP Target Asset Lists.ps1
Either SecurityCenter/Tenable.sc--or maybe the implementation we're using--does not succeed at scanning all systems for SCAP scanning when done in scans that are large. Large is a variable term here, but around 600, perhaps a bit more, seems viable. Larger scans will fail to build the results (though the scan does "complete").

As such, this script uses pre-created asset lists of a given naming scheme, and then updates the contents of the DNS list with information from Active Directory. It's not as simple and automated as I would like, but it does allow the scans to run.

Depending on the size of the OU being scanned, the number of chunks may need to be altered to get to a size that the SecurityCenter/Tenable.sc server can manage.