Create Get AD Object Last Logon.ps1
An example of what _not_ to do, but it was illustrative in arriving at the correct solution, so I am committing it anyway.

While it technically works to get a list of all objects from AD, then For-Each query each DC for the age of each object, it is abysmally slow. Worse, PoSH v5 doesn't have threading, and even if it did it still wouldn't be the best method. (Each object on each DC has a logon date, so why not just directly query each DC for every object, with the desired properties being returned? And that's the next commit.)