Drastically improve AD query speed
There's probably a more efficient way of handling the generic List we're using as the datastore here... it could break if there are sufficient objects or DCs polled to get the correct date. Possibly pre-compute the needed value to store it all locally? Akin to querying AD directly for the desire object type, storing the resulting count, then storing the quantity of DCs to query, taking the product plus some slack, then using that as the max List size?

Regardless, this 'Just Works', for the moment. And it doesn't make me hate myself for the prior commit's poor performance in querying AD.

Kinda still in progress, kinda complete. Possibly add a user prompt for user/computer object? And still need to dump this to a file of some description, but it works when run interactively (then dumped via something akin to ``$datastore | ConvertTo-Csv -NoTypeInformation | Out-File foo.csv``, or the like).